get_vespadb_obs

Query the vespa-db api to get observations

Usage

get_vespadb_obs(..., auth = NULL, domain = c("production", "uat"))

Arguments

...

<dynamic-dots> Query parameters according to the observation model.

auth

Character. Path to the authentication cookie to use. By default a request without authentication is placed (public data only). The authentication cookie can be created using login_vespadb() which returns the path to the cookie.

domain

Character. The domain to login to. Default is "uat" for the UAT environment.

Value

A tibble with the observations from the VespaDB API.

Details

Extra query parameters are supported according to the Observation model. More information on the allowed parameters can be fetched via get_field_definitions().

This function already handles paging, so supplying a page as a parameter is not allowed. For a more low level function to fetch results see fetch_result().

Examples

if (FALSE) { # \dontrun{
# All public records
get_vespadb_obs()

# Public records from Ghent
get_vespadb_obs(municipality_id = 215)

# Public and private information from Ghent, only ANB plots
get_vespadb_obs(
  municipality_id = 215,
  anb = "true",
  auth = login_vespadb()
)
} # }