Get the municipality ids.

This function either returns a complete overview of all municipality ids. Or if one or more municipalities are passed, the id's for these municipalities.

Usage

municipalities(municipalities = NULL, domain = c("production", "uat"))

Arguments

municipalities

(Optional) A character vector of municipality names. If this is not provided, all municipalities will be returned as a tibble with their ids.

domain

Character. The domain to login to. Default is "uat" for the UAT environment.

Value

Either a numeric vector of ids, or if no argument is passed, a tibble with all municipalities and their id.

Examples

municipalities("Gent")
#> [1] 215
municipalities(c("Gent", "Kortrijk"))
#> [1] 215 255
municipalities()
#> # A tibble: 285 × 2
#>    municipality    id
#>    <chr>        <int>
#>  1 Aalst          138
#>  2 Aalter         195
#>  3 Aarschot        95
#>  4 Aartselaar       4
#>  5 Affligem        27
#>  6 Alken          116
#>  7 Alveringem     294
#>  8 Antwerpen      271
#>  9 Anzegem        126
#> 10 Ardooie         10
#> # ℹ 275 more rows